Re-organized some classes.
