+++ server/src/main/java/edu/colostate/cs/cs414/chessirecoders/game/GameHistory.java
@@ -1,0 +1,4 @@
+package edu.colostate.cs.cs414.chessirecoders.game;
+
+public class GameHistory {
+}
