+++ server/src/main/java/edu/colostate/cs/cs414/chessirecoders/loginManager/LoginManager.java
@@ -1,0 +1,4 @@
+package edu.colostate.cs.cs414.chessirecoders.loginManager;
+
+public class LoginManager {
+}
