+++ server/src/main/java/edu/colostate/cs/cs414/chessirecoders/loginManager/LoginVerifier.java
@@ -1,0 +1,5 @@
+package edu.colostate.cs.cs414.chessirecoders.loginManager;
+
+public class LoginVerifier {
+
+}
