--- server/src/main/java/edu/colostate/cs/cs414/chessirecoders/server/GameHistory.java
@@ -1,4 +1,0 @@
-package edu.colostate.cs.cs414.chessirecoders.server;
-
-public class GameHistory {
-}
