--- server/src/main/java/edu/colostate/cs/cs414/chessirecoders/server/LoginVerifier.java
@@ -1,5 +1,0 @@
-package edu.colostate.cs.cs414.chessirecoders.server;
-
-public class LoginVerifier {
-
-}
